@@ -7,6 +7,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class MvcConfig implements WebMvcConfigurer {
     public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/user").setViewName("user");
+        registry.addViewController("/user").setViewName("user"); // связывает URL /user с Thymeleaf шаблоном user.html
+        // без необходимости создания отдельного контроллера
     }
 }
